WIP: UTIL_EST: switch to usage of tasks's estimated utilization
The task_util() call is used by the CFS scheduler to get a measure of
"how big a task is" in all the code paths where we need to decide
if a task can be allocated on a certain SG/CPU.
When the utilization estimation feature is enabled, we should always consider
the estimated utilization as a measure of the CPU bandwidth required
by that task.

This patch update all the task_util() calls to use the estimated utilization
metrics whenever the UTIL_EST sched feature is enabled.
That's should enforce a pervasive usage of a more stable signal to evaluate
task requirements in all sensible code paths.

Specifically, the patch switch to usage of estimated utilization:
a) for checks if a task fits into a given CPU,
   i.e. for the cluster selection in the energy aware wakeup path,
   via task_fits_max, and for load balancing paths, via task_fits_spare.
b) in the energy aware wakeup path
   for both the CPU selection and the energy model computation

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: Tab Fitts <tfitts1212@gmail.com>